Styles merged with React version
Styles have been aligned to version hosted here:
https://eliteness-solidly-table.vercel.app/

Minimal modifications to existing HTML generation routines so addition of further DEX instances should be trivial.

Steps:
- Add data feed to configuration
- Add addition CSS rule for colour variant td:n-child
- Add header and logo td

A version of CSS which is formatted and legible is included, but not referenced.

HTML refers to the not-quite-minified-css file.  Up to you if you want to make it all inline CSS.